requirements.txt
quart
pillow
